CTS integration
The sample app does its own function loading, and unconditionally
include Vulkan headers. For external integrations like the CTS, there
needs to be some hook points to control this loading and header
inclusion, so that we may ensure only the CTS versions are being
used.

This patch detects when the app is being built inside the CTS, and
arranges for the correct C++ namespace to be included so that the
sample app can be used as-is in the CTS, sharing more code.

A few minor compiler warnings have been fixed here too.