Add MESA AV1 vendor extension support
- Add video profiling handling
- Handle picture parameter updates

This is does not currently support parameter updates (More than one
set in the stream), since no driver handles them correctly yet.

- Tile parsing changes

Compared to FFmpeg, which creates bitstream buffers on demand
containing only the TILE_GROUP OBU + header, the NVIDIA parser uploads
bitstream buffers of 2MiB and fills them until they run out (and then
recycles using a pool). This means non-zero offset information must be
passed for NVIDIA for the first tile. It wasn't clear from the headers
what to offset to here. RADV works as described, but the spec should
make something clear, need to consult IHVs.

- Fixed bug in order hint parsing

The core fix is this,

-        frame_offset = u(sps->order_hint_bits_minus_1 + 1);
+        frame_offset = sps->flags.enable_order_hint ? u(sps->order_hint_bits_minus_1 + 1) : 0;

sps->order_hint_bits_minus_1 was originally in a bitfield, and -1 as
its default was OK (but very confusing to readers). In the MESA
headers, its an uint8_t, so -1 -> 255 and then +1 would then cause the
parser to read 256 bits, further pain ensued.

- Reimplemented tile parsing

The main issue fixed by this commit is that tile_height_in_sbs_minus_1
was not being correctly inferred by the original parsing code.

I also cross-checked the parsing code against aom & FFmpeg (which was
very similar the code the samples app)

The sample app is matching the FFmpeg pass rate now. The remaining
failures are to do with SVC and parameter set handling, plus the odd
weird case that FFmpeg is also failing (48delayed).