Remove needless indirection with ref frame maps
ref_frame_map was always an identify function. It's presence is
needlessly confusing the code.

active_ref_idx was just aliasing ref_frame_idx, and hence ref_frame[i]
was needless too.

In totality, this reduces a lot the amount of state to reason about
when reading the reference frame indexing code.